refactor(agent): the five files Phase 4 rule 1 names (#606)
The last Phase 2 row. Rule 1 locks "core files (`run.ts`, `turn.ts`,
`tools.ts`, `effects.ts`, `state.ts`) contain no domain string literals",
and none of those files existed.

Four are renames: `runner` -> `run`, `run-state` -> `state`,
`tool-executor` -> `tools`, `policy-effects` -> `effects`.

`turn.ts` folds `turn-prepare` and `turn-outcome` together, which rule 2
asks for rather than merely permits: a split needs a second consumer, a
trust boundary, or independent lifecycles, and preparing a turn and
settling it have none of the three — both are reached only from the run
loop and both change when the turn's shape does. "The file is getting
long is not on this list — line count is a review signal, never a split
criterion."

`run-events.ts`, `lifecycle-dispatch.ts` and `compaction.ts` keep their
names. The first has many importers; the second is the run-level points,
which the rule does not name; the third moves in Phase 3.
