+++ packages/agent/src/core/execution/run.ts
@@ -4,8 +4,8 @@ import type { AgentResult, ChatAgentConfig, ChatAgentInput } from "../types";
 import * as Retry from "../retry";
 import { PolicyEngine, type PolicyEngineInstance } from "../policy";
 import { emitRunCompleted, emitRunFailed, emitRunStarted, emitTurnStart } from "./run-events";
-import { handleCompact, handleContinue, handleError, handleStop } from "./turn-outcome";
-import { assertToolExecutor, buildTurn, resolveToolChoice } from "./turn-prepare";
+import { handleCompact, handleContinue, handleError, handleStop } from "./turn";
+import { assertToolExecutor, buildTurn, resolveToolChoice } from "./turn";
 import {
   dispatchBudgetCheck,
   dispatchModelRequest,
@@ -18,7 +18,7 @@ import {
   requireTrace,
   type AgentRunBase,
   type RunFailureFacts,
-} from "./run-state";
+} from "./state";
 
 /**
  * Runs an agent to a result.
