+++ packages/agent/test/core/execution/lifecycle-error.test.ts
@@ -4,7 +4,7 @@ import { Bus } from "@openomni/telemetry";
 import { PolicyEngine } from "../../../src/core/policy";
 import type { PolicyContext } from "../../../src/core/policy/types";
 import { abortRun, allow } from "../../helpers/policy-decision";
-import { handleError } from "../../../src/core/execution/turn-outcome";
+import { handleError } from "../../../src/core/execution/turn";
 import { makeAgentBase, makeConfig, makeState } from "./lifecycle-dispatch-fixture";
 
 describe("handleError (error)", () => {
