+++ packages/agent/test/core/execution/tool-executor-sole-emitter.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "bun:test";
 import { ToolExecution } from "@openomni/protocol";
 import type { Tool } from "@openomni/protocol";
 import { Bus } from "@openomni/telemetry";
-import { createToolExecutor } from "../../../src/core/execution/tool-executor";
+import { createToolExecutor } from "../../../src/core/execution/tools";
 import { PolicyEngine } from "../../../src/core/policy";
 
 // #522 defect 2 — sole-emitter pin. The worker-side executor
