+++ packages/agent/test/core/policy/canonical-execution.test.ts
@@ -3,9 +3,9 @@ import type { CanonicalAuditDispatchContextGeneric } from "@openomni/policy";
 import type { CanonicalPolicyRegistrationGeneric } from "@openomni/policy";
 import { Operational, PolicyDecision } from "@openomni/protocol";
 import { Bus } from "@openomni/telemetry";
-import { createToolExecutor } from "../../../src/core/execution/tool-executor";
-import { buildPolicyEngine } from "../../../src/core/execution/runner";
-import { runAgent } from "../../../src/core/execution/runner";
+import { createToolExecutor } from "../../../src/core/execution/tools";
+import { buildPolicyEngine } from "../../../src/core/execution/run";
+import { runAgent } from "../../../src/core/execution/run";
 import { createCompactionPolicy } from "../../../src/core/policy/builtin";
 import type { PolicyContext } from "../../../src/core/policy/types";
 import type { ChatAgentConfig } from "../../../src/core/types";
