+++ packages/agent/test/core/policy/native-tool-post-dispatch.test.ts
@@ -1,7 +1,7 @@
 import { expect, it } from "bun:test";
 import { PolicyDecision, type Tool } from "@openomni/protocol";
 import { Bus } from "@openomni/telemetry";
-import { createToolExecutor } from "../../../src/core/execution/tool-executor";
+import { createToolExecutor } from "../../../src/core/execution/tools";
 import { PolicyEngine } from "../../../src/core/policy";
 
 /**
